+++ ui/app/page.tsx
@@ -1,5 +1,4 @@
 import Logout from '@/components/Logout';
-import Password from '@/components/Password';
 import SidebarComponent from '@/components/SidebarComponent';
 import { Header } from '@/lib/Header';
 import { Layout, LayoutMain } from '@/lib/Layout';
